This PR is the companion of https://github.com/JuliaGeometry/Quaternions.jl/pull/56. Like that PR for quaternions, it draws on a more general result that allows us to extend all complex analytic functions in Base to the octonions with a function `extend_analytic`. 

Since `extend_analytic` is faster than the current implementations of `sqrt` and `exp`, those are refactored to use it. The PR also speeds up `log` and, for clarity, refactors it to adopt the same notation as `extend_analytic`.

Here's a performance comparison:

```julia
julia> using Octonions, BenchmarkTools

julia> o = randn(OctonionF64);
```

Before:
```julia
julia> @btime exp($o);
  54.570 ns (0 allocations: 0 bytes)

julia> @btime sqrt($o);
  184.861 ns (0 allocations: 0 bytes)

julia> @btime log($o);
  99.939 ns (0 allocations: 0 bytes)
```

After:
```julia
julia> @btime exp($o);
  42.833 ns (0 allocations: 0 bytes)

julia> @btime sqrt($o);
  38.143 ns (0 allocations: 0 bytes)

julia> @btime log($o);
  67.569 ns (0 allocations: 0 bytes)
```